backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0
PyPDF2==3.0.1
pydantic==2.5.0
aiofiles==23.2.1
